quote μs to allow using in the browser
Quotes non-ascii `μ` character to allow use in the browser.

Having this character in strings is fine, but including it in
variable/property names requires everyone using any code that
contains this module in a final bundle to add `charset='utf-8'`
to their script tags.

With this change they can continue omitting the charset.